requirement.txt
ipykernel==6.30.1
medpy==0.5.2
monai==1.5.0
tensorboard==2.20.0
nibabel==5.3.2
pandas==2.3.1
pytorch_lightning==2.5.3
thop==0.1.1.post2209072238
torch_pruning==1.6.0
numpy==1.26.4
einops==0.8.1